+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/StubExoPlayer.java
@@ -412,4 +412,9 @@ public class StubExoPlayer extends StubPlayer implements ExoPlayer {
   public boolean experimentalIsSleepingForOffload() {
     throw new UnsupportedOperationException();
   }
+
+  @Override
+  public boolean isTunnelingEnabled() {
+    throw new UnsupportedOperationException();
+  }
 }